+++ src/main.rs
@@ -118,7 +118,7 @@ impl Drop for DashPassthroughGuard {
 fn main() {
     let size = TerminalSize::current();
     eprintln!(
-        "dashcompositor v0.8.0 -- multi-layer + auto-detect encoder: \
+        "dashcompositor v0.8.2 -- multi-layer + auto-detect encoder: \
 host terminal = {cols} cols x {rows} rows",
         cols = size.cols,
         rows = size.rows,